DRIVER NOTIFICATION BOT/requirements.txt
aiogram>=3.2.0,<4.0.0
asyncpg>=0.29.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
